make AnnAssign Py 3.6+
